requirements.txt
google-cloud-texttospeech>=2.0.0,<3.0.0
gTTS>=2.5.0,<3.0.0
